Rescaled relative position in motif
The position within a motif was rescaled to be from -1 to 1, with 0 marking the motif center. The
regression analysis was updated to use the absolute value of that position. In those models, a value
of positive 1 indicates the edge of the motif, and a value of 0 indicates the center of the motif model.